+++ packages/typespec-python/CHANGELOG.md
@@ -20,6 +20,7 @@
 
 - Change readonly to visibility #1968
 - Support global config for `head-as-boolean` #1949
+- Support global config for `skip-special-headers` #1973
 
 ## 2023-06-12 - 0.11.0